Function Documentations

Functions from file example/BEA2016...-input/func_run_example.jl

LWFBrook90.run_example — Method

run_example()

Run example simulation located in "/example/BEA2016-reset-FALSE-input" for 100 days and return a Dict containing the solution (solution object of DifferentialEquations.jl) and other variables useful for plotting.

Example:

using LWFBrook90
using Plots
example = LWFBrook90.run_example()

# Plot scalar solution
# Using dates (but not interpolated)
plot(example["solutionDates"],
    example["solution"][[1,2,3,4,5,6],:]',
    label=["GWAT" "INTS" "INTR" "SNOW" "CC" "SNOWLQ"])

# Using simple plot recipe that interpolates, but without dates
plot(example["solution"];
    vars = [1, 2, 3, 4, 5, 6],
    label=["GWAT" "INTS" "INTR" "SNOW" "CC" "SNOWLQ"])

# Plot vector solution
x = example["solutionDates"]
y = cumsum(example["thickness"])
z = example["solution"][7 .+ (0:example["NLAYER"]-1), :]./example["thickness"]
heatmap(x, y, z,
    yflip = true,
    xlabel = "Date",
    ylabel = "Depth",
    colorbar_title = "θ")

Functions from file LWFBrook90.jl

Functions from file func_read_inputData.jl

LWFBrook90.DateTime2RelativeDaysFloat — Method

DateTime2RelativeDaysFloat(x,reference_DateTime)

Transforms DateTimes x to simulation time

LWFBrook90.RelativeDaysFloat2DateTime — Method

RelativeDaysFloat2DateTime(t, reference_DateTime)

Transforms simulation time t to DateTimes

LWFBrook90.p_DOY — Method

p_DOY(t::Float64, reference::DateTime)

Get DOY (Day Of Year) from simulation time

LWFBrook90.p_MONTHN — Method

p_MONTHN(t::Float64, reference::DateTime)

Get Month from simulation time

LWFBrook90.read_LWFBrook90R_inputData — Method

read_LWFBrook90R_inputData(folder::String, prefix::String)

Load different input files for LWFBrook90:

meteoveg

param

siteparam

pdur

soil_materials.csv

soil_nodes.csv

These files were created with an R script generate_LWFBrook90Julia_Input.R that takes the same arguements as the R funciton LWFBrook90R::run_LWFB90() and generates the corresponding Julia input functions.

Functions defining the DiffEq.jl system of ODE (p, u0, f, callbacks, ...)

LWFBrook90.define_LWFB90_u0 — Method

define_LWFB90_u0()

Generate vector u0 needed for ODE() problem in DiffEq.jl package.

LWFBrook90.define_LWFB90_p — Method

define_diff_eq_parameters()

Generate vector p needed for ODE() problem in DiffEq.jl package.

Arguments

NLAYER::...: TODO argument description.

IMODEL::...: TODO argument description.

constant_dt_solver::...: TODO argument description.

NOOUTF::...: TODO argument description.

Reset::...: TODO argument description.

compute_intermediate_quantities::...: TODO argument description.

input_meteoveg::...: TODO argument description.

input_siteparam::...: TODO argument description.

input_param::...: TODO argument description.

input_soil::...: TODO argument description.

input_pdur::...: TODO argument description.

LWFBrook90.discretize_soil_params — Method

discretize_soil_params(input_soil_materials,IMODEL, ILAYER, QLAYER, NLAYER)

Define constant parameters from input_pdur. TO BE REDEFINED

LWFBrook90.interpolate_meteoveg — Method

interpolate_meteoveg(input_meteoveg::DataFrame, input_meteoveg_reference_date::DateTime)

Take climate and vegetation parameters in input_meteoveg and generates continuous parameters.

LWFBrook90.define_LWFB90_cb — Method

define_LWFB90_cb()

Generate callback function cb needed for ODE() problem in DiffEq.jl package.

LWFBrook90 updates states INTS, INTR, SNOW, CC, SNOWLQ not continuously but only once per day. This operator splitting (daily vs continuous update of ODEs) is implemented by using this callback function which is called once per day.

LWFBrook90.define_LWFB90_ODE — Method

define_LWFB90_ODE()

Generates an ODEProblem from DiffEq.jl

An ODE problem which consists of

definition of right-hand-side (RHS) function f

definition of callback function cb

initial condition of states

definition of simulation time span

parameters

Seperate updating of different states (INTS, INTR, SNOW, CC, SNOWLQ are updated once per day while GWAT and SWATI are updated continuously) is implemented by means of operator splitting using a callback function for the daily updates and a ODE RHS (right hand side) for the continuous update.

LWFBrook90.MSBSETVARS — Method

MSBSETVARS()

Compute state dependent parameters for updating states INTS, INTR, SNOW, CC, SNOWLQ in callback function.

Arguments

many

Functions from the different modules defining LWF-BROOK90

Modules = [LWFBrook90.CONSTANTS,
           LWFBrook90.GLBLDECL,
           LWFBrook90.KPT,
           LWFBrook90.WAT,
           LWFBrook90.SUN,
           LWFBrook90.PET,
           LWFBrook90.SNO,
           LWFBrook90.EVP
           ]
Pages   = ["module_CONSTANTS.jl",
           "module_GLBLDECL.jl",
           "module_KPT.jl",
           "module_WAT.jl",
           "module_SUN.jl",
           "module_PET.jl",
           "module_SNO.jl",
           "module_EVP.jl"
           ]
Order = [:constant, :type, :function, :macro]